+++ include/libtimestep/rotational_system/rotational_unary_system.h
@@ -30,7 +30,7 @@ class rotational_unary_system : public rotational_generic_system<field_value_t, 
         rotational_unary_system<field_value_t, real_t, integrator_t, step_handler_t, acceleration_handler_t>> {
 public:
     typedef std::vector<field_value_t> field_container_t;
-    typedef std::vector<size_t> index_container_t;
+    typedef std::vector<long> index_container_t;
 
     // Class constructor
     //
