+++ include/libtimestep/step_handler/step_handler.h
@@ -16,7 +16,7 @@ template <typename field_container_t, typename field_value_t>
 struct step_handler {
 
     // This method increments the specified value in the x buffer
-    void increment_x(size_t n,                                                                          // index of the value to increment
+    void increment_x(long n,                                                                          // index of the value to increment
                     field_value_t const & dx,                                                           // value of the position increment
                     typename field_container_t::iterator x_begin_itr,                                   // iterator pointing to the start of the x buffer
                     typename field_container_t::const_iterator v_begin_itr [[maybe_unused]],            // iterator pointing to the start of the v buffer
@@ -26,7 +26,7 @@ struct step_handler {
     }
 
     // This method increments the specified value in the v buffer
-    void increment_v(size_t n,                                                                          // index of the value to increment
+    void increment_v(long n,                                                                          // index of the value to increment
                      field_value_t const & dv,                                                          // value of the velocity increment
                      typename field_container_t::const_iterator x_begin_itr [[maybe_unused]],           // iterator pointing to the start of the x buffer
                      typename field_container_t::iterator v_begin_itr,                                  // iterator pointing to the start of the v buffer
