+++ include/libtimestep/system/unary_system.h
@@ -30,7 +30,7 @@ class unary_system : public generic_system<field_value_t, real_t, integrator_t, 
         unary_system<field_value_t, real_t, integrator_t, step_handler_t, acceleration_handler_t>> {
 public:
     typedef std::vector<field_value_t> field_container_t;
-    typedef std::vector<size_t> index_container_t;
+    typedef std::vector<long> index_container_t;
 
     // Class constructor
     //
@@ -58,7 +58,7 @@ public:
 
         this->reset_acceleration_buffer();
 
-        std::for_each(this->indices.begin(), this->indices.end(), [t, this] (size_t i) {
+        std::for_each(this->indices.begin(), this->indices.end(), [t, this] (long i) {
             this->a[i] += acceleration_handler.compute_acceleration(i, this->get_x(), this->get_v(), t);
         });
     }
